Work around GitHub new-file directory bug
It used to be that providing a `filename` parameter would make it so
that you are creating a file with that name in the specified directory.
At some point, GitHub seems to have introduced a bug that drops the last
directory.  Adding an underscore makes the target directory what we
want.